feat(kitty): split neo-tree launcher
